"move location of optimizer library"
