+++ paddle/optimizer/adagrad_optimizer.cc
@@ -1,0 +1,4 @@
+#include "sgd_optimizer.h"
+
+namespace paddle {
+namespace optimizer {
